address pr feedback
- use process oriented examples
