+++ bindings/rust-examples/client-hello-config-resolution/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "client-hello-config-resolution"
+version.workspace = true
+authors.workspace = true
+publish.workspace = true
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+clap = { version = "4", features = ["derive"] }
+s2n-tls = { version = "0.1", path = "../../rust/s2n-tls" }
+s2n-tls-tokio = { version = "0.1", path = "../../rust/s2n-tls-tokio" }
+tokio = { version = "1", features = ["full"] }
