+++ bindings/rust-examples/client-hello-config-resolution/certs/ca-cert.pem
