+++ bindings/rust-examples/client-hello-config-resolution/certs/generate.sh
