+++ bindings/rust-examples/client-hello-config-resolution/certs/kangaroo-key.pem
