--- bindings/rust/examples/client-hello-config-resolution/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "client-hello-config-resolution"
-version = "0.0.1"
-authors = ["AWS s2n"]
-edition = "2021"
-
-[dependencies]
-s2n-tls = { version = "0.1", path = "../../s2n-tls" }
-s2n-tls-tokio = { version = "0.1", path = "../../s2n-tls-tokio" }
-tokio = { version = "1", features = ["full"] }
-turmoil = "0.6"
