--- bindings/rust/examples/client-hello-config-resolution/README.md
@@ -1,1 +1,0 @@
-This example shows how to use the s2n-tls client hello callback to configure different TLS configs based on the Server Name Indication (SNI) in the client hello. To make running the example easier, the example is run using [Turmoil](https://github.com/tokio-rs/turmoil) which can function as a network simulator. This allows us to simulate two different clients and a single server without having to spin up mupltiple processes.
